separated app and index, fixed test, added test util
