+++ backend/src/index.ts
@@ -1,27 +1,8 @@
-// Import modules from node_modules
-import * as express from 'express';
-import * as bodyParser from 'body-parser';
-import * as cors from 'cors';
+import app from "./app";
+import { AppDataSource } from "./data-source";
 
-// Import from other .ts files
-import { AppDataSource } from './data-source';
-import { resolve } from 'path';
-import { api } from './routes/index';
+// Initalizes data source
+AppDataSource.initialize();
 
-// Create express app
-const app = express();
-
-AppDataSource.initialize().then(async () => {
-  // Using third party middleware
-  app.use(bodyParser.json()); // body-parser is which allows express to read the body and then parse that into a Json object that we can understand.
-  app.use(cors()); //
-
-  // Routes go here
-  // Creating route : https://expressjs.com/en/guide/routing.html
-  app.use('/api', api);
-
-  // Starts server and listens on port 3001 for connections
-  app.listen(3001);
-});
-
-export default app;
+// Express app listens to port 3001
+app.listen(3001);
